==> 1_main/3_plots_corrs.Rmd <==
---
title: "3_plot_corr"
output: html_document
date: "2024-02-19"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(deSolve)
library(ggplot2)
library(tidyr)
library(ggpubr)
library(ggmatplot)
library(RColorBrewer)
library(gridExtra)
library(cowplot)
library(viridis)
```


```{r}
tot_pop <- 30066648   ## Took population for 2020 from Mozambique INE (Institute of statistics)
p_urban <- 0.34 ##From INE
p_rural <- 1-p_urban
start.Ns <- c(10884513, 4883969, 7958844, 4900719, 992149, 446454)
dist <- start.Ns/tot_pop

pop_dist = data.frame(age_ur = c("cr","cu","ar","au","er","eu"),
                      pop = start.Ns)

old_pop <-  992149+446454
```


```{r}
inc_time <- data.frame(sn = seq(from=0,to=9, by=1),
                       wave= seq(from=1,to=10, by=1))%>%
  mutate(time= sn*365+199)

age_dist <- data.frame(age = c("c","a","e"),
                       pop = c(10884513+4883969, 7958844+4900719, 992149+446454))

sweep<- readRDS("0_sweep_imm4_foisp0.6.RDS")%>%
  mutate(sero_thresh=as.character(sero_thresh))%>%
  slice(1:384)


imm_grid <- data.frame(
  v0=c(1,               sweep$rel_1v0[1],sweep$rel_2v0[1]),
  v1=c(sweep$rel_0v1[1],sweep$rel_1v1[1],sweep$rel_2v1[1]),
  v2=c(sweep$rel_0v2[1],sweep$rel_1v2[1],sweep$rel_2v2[1]),
  v3=c(sweep$rel_0v3[1],sweep$rel_1v3[1],sweep$rel_2v3[1]))
row.names(imm_grid)=c("exp0","exp1","exp2")

imm_vec <- imm_grid%>%
  mutate(exp = row.names(.))%>%
  pivot_longer(cols=v0:v3, names_to="vax",values_to = "val")%>%
  arrange(vax,exp)

imm_vec <- unlist(imm_vec$val)

```

##Read in novax
```{r}
deaths <- readRDS("0_res/vax_death.RDS")
sero   <- readRDS("0_res/vax_sero.RDS")
imm    <- readRDS("0_res/vax_imm.RDS")
inc    <- readRDS("0_res/vax_inc.RDS")

deaths <- deaths%>%filter(sero_thresh=="0")
sero   <- sero%>%filter(sero_thresh=="0")
imm    <- imm%>%filter(sero_thresh=="0")
inc    <- inc%>%filter(sero_thresh=="0")
```

```{r}
## immunity per wave


#sus_val <- data.frame(imm_cat = unique(imm$imm_cat),
#                      sus_val = c(1,0.54,0.23,0.62,0.33,0.14,0.46,0.25,0.1,0))

sus_val <- data.frame(imm_cat=unique(imm$imm_cat),
                      sus_val = c(imm_vec[1:3],imm_vec[7:12],0))

imm_wave <- imm %>% filter(time %in% unlist(inc_time$time))
imm_wave <- imm_wave%>% left_join(
  sus_val, by = c("imm_cat"="imm_cat")
)


imm_wave1 <- imm_wave%>%
  select(-sus_val, -var)%>%
  tidyr::pivot_wider(names_from = "imm_cat", values_from="val")%>% 
  unchop(everything())%>%
  replace_na(list(s0v2 = 0, s1v2=0, s2v2=0, s0v3=0, s1v3=0, s2v3=0))%>%
  mutate(
    sus_tot = s0v0+s1v0+s2v0+s0v2+s1v2+s2v2+s0v3+s1v3+s2v3) %>% 
  left_join(age_dist, by = c("age_grp"="age"))%>%
  mutate(tot = sus_tot+imm1)



imm_wave2<- imm_wave %>% mutate(
  sus_pop = val*sus_val
)%>%
  select(-sus_val, -val,-var)%>%
  replace_na(list(sus_pop = 0))%>%
  pivot_wider(names_from = "imm_cat", values_from="sus_pop")%>% 
  unchop(everything())%>%
  replace_na(list(s0v2 = 0, s1v2=0, s2v2=0, s0v3=0, s1v3=0, s2v3=0))%>%
  
  mutate(
    sus_tot = s0v0+s1v0+s2v0+s0v2+s1v2+s2v2+s0v3+s1v3+s2v3) %>% 
  select(-imm1)%>%
  left_join(imm_wave1%>%select(time,sweep_unique,sero_thresh,age_grp, tot, imm1, omega,kappa,foi_sp), 
            by = c("age_grp"="age_grp","time"="time",
                   "sweep_unique"="sweep_unique","sero_thresh"="sero_thresh", 
                   "omega"="omega","kappa"="kappa", "foi_sp"="foi_sp")) %>% 
  mutate(prop_sus = sus_tot/tot,
         prop_imm = imm1/tot)



imm_wave_simp <- imm_wave2%>%
  select(time, sweep_unique, sero_thresh, omega,kappa, age_grp, prop_sus, prop_imm,foi_sp)%>%
  pivot_wider(names_from = "age_grp", values_from = c("prop_sus","prop_imm"))%>%
  unchop(everything())%>%
  left_join(inc_time)
prop_sus_long <- imm_wave_simp%>%select(wave, sweep_unique, sero_thresh,omega, kappa,foi_sp, prop_sus_c:prop_sus_e)%>%
  pivot_longer(cols=prop_sus_c:prop_sus_e, names_to = "age",values_to="prop_sus")%>%
  unchop(everything())%>%
  mutate(age = substr(age,10,10))

prop_imm_long <- imm_wave_simp%>%select(wave, sweep_unique, sero_thresh,omega,kappa, foi_sp,prop_imm_c:prop_imm_e)%>%
  pivot_longer(cols=prop_imm_c:prop_imm_e, names_to = "age",values_to="prop_imm")%>%
  unchop(everything())%>%
  mutate(age = substr(age,10,10))
```


```{r}
#####################
###Make df_corr###
##################



## deaths per wave
death_wave <- deaths%>%
  mutate(omega = paste0(omega2,"-",omega4))%>%
  mutate(
    wave = case_when(time<inc_time$time[1]~NA_real_,
                     time<inc_time$time[2]~1,
                     time<inc_time$time[3]~2,
                     time<inc_time$time[4]~3,
                     time<inc_time$time[5]~4,
                     time<inc_time$time[6]~5,
                     time<inc_time$time[7]~6,
                     time<inc_time$time[8]~7,
                     time<inc_time$time[9]~8,
                     time<inc_time$time[10]~9,
                     TRUE~10)
  )%>%
  group_by(sweep_unique, sero_thresh, wave, omega,kappa, foi_sp)%>%
  dplyr::summarise(c = sum(new_Deaths_c),
                   a = sum(new_Deaths_a),
                   e = sum(new_Deaths_e),
                   all=sum(new_Deaths_tot)
  )
death_wave_long<- death_wave%>%pivot_longer(cols=c:all, names_to="age", values_to = "tot_death")


##seroprev per wave
sero_wave <- sero%>%
  mutate(omega = paste0(omega2,"-",omega4))%>%
  mutate(
    wave = case_when(time<inc_time$time[1]~NA_real_,
                     time<inc_time$time[2]~1,
                     time<inc_time$time[3]~2,
                     time<inc_time$time[4]~3,
                     time<inc_time$time[5]~4,
                     time<inc_time$time[6]~5,
                     time<inc_time$time[7]~6,
                     time<inc_time$time[8]~7,
                     time<inc_time$time[9]~8,
                     time<inc_time$time[10]~9,
                     TRUE~10)
  )%>%
  group_by(sweep_unique, sero_thresh, wave, omega,kappa, foi_sp)%>%
  slice(1)

sero_wave_long <- sero_wave%>%pivot_longer(cols=c(sero_c:sero_e,sero_all), names_to="age",values_to = "sero")%>%
  mutate(age = gsub("sero_","",age))




#saveRDS(prop_sus_long, "prop_sus_long.RDS")
#saveRDS(prop_imm_long, "prop_imm_long.RDS")

## deaths per wave
inc_wave <- inc%>%
  mutate(omega = paste0(omega2,"-",omega4))%>%
  mutate(
    wave = case_when(time<inc_time$time[1]~NA_real_,
                     time<inc_time$time[2]~1,
                     time<inc_time$time[3]~2,
                     time<inc_time$time[4]~3,
                     time<inc_time$time[5]~4,
                     time<inc_time$time[6]~5,
                     time<inc_time$time[7]~6,
                     time<inc_time$time[8]~7,
                     time<inc_time$time[9]~8,
                     time<inc_time$time[10]~9,
                     TRUE~10)
  )%>%
  group_by(sweep_unique, sero_thresh, wave, omega,kappa, foi_sp)%>%
  dplyr::summarise(inc_t = sum(inc_all),
                   inc_c   = sum(inc_c),
                   inc_a   = sum(inc_a),
                   inc_e   = sum(inc_e))

inc_wave_long <- inc_wave%>%
  select(wave, sweep_unique, sero_thresh,omega,kappa, foi_sp, inc_t:inc_e)%>%
  pivot_longer(cols=inc_t:inc_e, names_to = "age",values_to="cases")%>%
  unchop(everything())%>%
  mutate(age = substr(age,5,5))%>%
  mutate(age = ifelse(age=="t","all",age))




##Combine all the above by wave
df_corr_long <- death_wave_long%>%
  left_join(
    sero_wave_long%>%select(-time, -date), 
    by = c("wave"="wave","sweep_unique"="sweep_unique",
           "sero_thresh"="sero_thresh", "age"="age", "omega"="omega","kappa"="kappa",
           "foi_sp"="foi_sp")
  )%>%
  left_join(
    prop_sus_long,
    by = c("wave"="wave","sweep_unique"="sweep_unique",
           "sero_thresh"="sero_thresh", "age"="age", "omega"="omega","kappa"="kappa",
           "foi_sp"="foi_sp")
  )%>%
  left_join(
    prop_imm_long,
    by = c("wave"="wave","sweep_unique"="sweep_unique",
           "sero_thresh"="sero_thresh", "age"="age", "omega"="omega","kappa"="kappa",
           "foi_sp"="foi_sp")
  )%>%
  left_join(
    inc_wave_long,
    by = c("wave"="wave","sweep_unique"="sweep_unique",
           "sero_thresh"="sero_thresh", "age"="age", "omega"="omega","kappa"="kappa",
           "foi_sp"="foi_sp")
  )%>%
  mutate(Year=wave)

df_corr<- df_corr_long%>% filter(!is.na(wave))%>%
  mutate(Year=as.factor(Year),
         omega =paste0(omega2,"-",omega4))

##Correlations
df_corr <- df_corr%>%
  mutate(age_lab = 
           case_when(age=="a"~"Adult",
                     age=="e"~"Older adult",
                     age=="c"~"Child"))%>%
  mutate(age_lab = factor(age_lab, levels=c("Child","Adult","Older adult")),
         kappa = paste0("1","/",kappa),
         omega_lab = case_when(
           omega2==180~"Fast",
           omega2==270~"Medium",
           omega2==365~"Slow"))

corr_tab<- df_corr%>% filter(age!="all")%>%
  filter(wave>0)%>%
  filter(sero_thresh==0)%>%
  
  group_by(omega_lab, kappa, age_lab,age, foi_sp)%>%
  dplyr::summarise(sero_death = round(cor(sero,tot_death, method = "spearman", use="complete.obs"), digits=3),
                   sero_imm  =  round(cor(sero,prop_imm, method = "spearman", use="complete.obs"), digits=3),
                   sero_sus  =  round(cor(sero,prop_sus, method = "spearman", use="complete.obs"), digits=3),
                   sus_death =   round(cor(prop_sus,tot_death,method = "spearman", use="complete.obs"), digits=3),
                   sero_cases = round(cor(sero,cases, method = "spearman", use="complete.obs"), digits=3))


#Visualizing correlations

p1a <- df_corr%>%filter(wave>0 & sero_thresh==0&age=="e")%>%
  #filter(!kappa %in% c("1/1000","1/2000","1/3000"))%>%
  mutate(Year = as.factor(Year))%>%
  
  ggplot(aes(x=sero, y=tot_death, col = as.factor(Year)))+
  geom_point()+ xlab("Seroprevalence")+ylab("Deaths")+
  scale_color_brewer(palette="RdPu", direction=-1, name="Year")+
  theme_bw()+
  facet_grid(vars(kappa),vars(foi_sp))


p1b <-  df_corr%>%filter(wave>0 & sero_thresh==0&age=="e")%>%
  #filter(!kappa %in% c("1/1000","1/2000","1/3000"))%>%
  ggplot(aes(x=sero,y=prop_imm, col = Year))+
  geom_point()+ xlab("Seroprevalence")+ylab("Prop immune")+
  scale_color_brewer(palette="YlGn", direction=-1, name="Year")+
  theme_bw()+
  facet_grid(vars(kappa),vars(foi_sp))

p1c <-  df_corr%>%filter(wave>0 & sero_thresh==0&age=="e")%>%
  filter(!kappa %in% c("1/1000","1/2000","1/3000"))%>%
  ggplot(aes(x=sero, y=prop_sus, col = Year))+
  geom_point()+ xlab("Seroprevalence")+ylab("Relative susceptibility")+
  scale_color_brewer(palette="BuPu", direction=-1, name="Year")+
  theme_bw()+
  facet_grid(vars(kappa),vars(foi_sp))

p1d <- df_corr%>%filter(wave>0 & sero_thresh==0&age=="e")%>%
  filter(!kappa %in% c("1/1000","1/2000","1/3000"))%>%
  ggplot(aes(x=prop_sus, y=tot_death, col = Year))+
  geom_point()+ 
  scale_color_brewer(palette="BuGn", direction=-1, name="Year")+
  theme_bw()+xlab("Proportion susceptible")+ylab("Total deaths")+
  facet_wrap(~omega_lab)

df_corr%>%filter(wave>0 & sero_thresh==0&age=="e")%>%
  #filter(!kappa %in% c("1/1000","1/2000","1/3000"))%>%
  mutate(Year = as.factor(Year))%>%
  
  ggplot(aes(x=sero, y=cases, col = as.factor(Year)))+
  geom_point()+ xlab("Seroprevalence")+ylab("Deaths")+
  scale_color_brewer(palette="RdPu", direction=-1, name="Year")+
  theme_bw()

###
corr_sero_sus <- corr_tab%>%
  ggplot(aes(x=foi_sp, y=kappa))+
  geom_tile(aes(fill=sero_sus))+
  geom_text(aes(label=sero_sus))+
  facet_wrap(~age_lab)+
  scale_fill_gradient2(low ="#D73027", high = "#273871", mid = "white", midpoint = 0)+
  ggtitle("Correlation between seroprevalence and susceptibility at start of waves")+
  xlab("FOI_sp")+ylab("Rate of Waning antibody")

corr_sero_death <- corr_tab%>%
  ggplot(aes(x=foi_sp, y=kappa))+
  geom_tile(aes(fill=sero_death))+
  geom_text(aes(label=sero_death))+
  facet_wrap(~age_lab)+
  scale_fill_gradient2(low ="#D73027", high = "#273871", mid = "white", midpoint = 0)+
  ggtitle("Correlation between seroprevalence at start of waves and death")+
  xlab("FOI_sp")+ylab("Rate of Waning antibody")


corr_sero_case<- corr_tab%>%
  ggplot(aes(x=foi_sp, y=kappa))+
  geom_tile(aes(fill=sero_cases))+
  geom_text(aes(label=sero_cases))+
  facet_wrap(~age_lab)+
  scale_fill_gradient2(low ="#D73027", high = "#273871", mid = "white", midpoint = 0)+
  ggtitle("Correlation between seroprevalence at start of waves and cases")+
  xlab("FOI_sp")+ylab("Rate of Waning antibody")

p1a
p1b
p1c
p1d
```
